+++ src/test/helper/AppClassWithBindings.php
@@ -71,5 +71,5 @@ class AppClassWithBindings extends App
     /**
      * runs the command
      */
-    public function run() { }
+    public function run(): void { }
 }