+++ src/test/helper/AppClassWithInvalidBindingModule.php
@@ -34,5 +34,5 @@ class AppClassWithInvalidBindingModule extends App
     /**
      * runs the command
      */
-    public function run() { }
+    public function run(): void { }
 }