+++ src/test/helper/AppTestBindingModuleOne.php
@@ -20,7 +20,7 @@ class AppTestBindingModuleOne implements BindingModule
      * @param  \stubbles\ioc\Binder  $binder
      * @param  string                $projectPath  optional  project base path
      */
-    public function configure(Binder $binder, string $projectPath = null)
+    public function configure(Binder $binder, string $projectPath = null): void
     {
         $binder->bind('foo')->to('\\stdClass');
     }