+++ src/test/helper/AppUsingBindingModule.php
@@ -54,5 +54,5 @@ class AppUsingBindingModule extends App
     /**
      * runs the command
      */
-    public function run() { }
+    public function run(): void { }
 }